Add preliminary top level README.md
The README still need to have some links, pictures, and maybe formatting
added, but the majority of the content for the front page should be
there. A number of other pages need to be written as well, including
installation, getting started, and best practices, as well as
information about the labeling and training setup (and best practices
there too).